Cleanup editor route
